Solve #1018, Fill backup info for copy related rules (#1029)

+++ smart-common/src/main/java/org/smartdata/metaservice/BackupMetaService.java
@@ -25,11 +25,11 @@ public interface BackupMetaService extends MetaService {
 
   List<BackUpInfo> listAllBackUpInfo() throws MetaServiceException;
 
-  BackUpInfo getBackUpInfoById(int id) throws MetaServiceException;
+  BackUpInfo getBackUpInfoById(long id) throws MetaServiceException;
 
   void deleteAllBackUpInfo() throws MetaServiceException;
 
-  void deleteBackUpInfoById(int id) throws MetaServiceException;
+  void deleteBackUpInfoById(long id) throws MetaServiceException;
 
   void insertBackUpInfo(BackUpInfo backUpInfo) throws MetaServiceException;
