+++ smart-metastore/src/main/java/org/smartdata/metastore/MetaStore.java
@@ -1134,7 +1134,7 @@ public class MetaStore implements CopyMetaService, CmdletMetaService, BackupMeta
     }
   }
 
-  public BackUpInfo getBackUpInfoById(int id) throws MetaStoreException {
+  public BackUpInfo getBackUpInfoById(long id) throws MetaStoreException {
     try {
       return backUpInfoDao.getById(id);
     } catch (Exception e) {
@@ -1150,7 +1150,7 @@ public class MetaStore implements CopyMetaService, CmdletMetaService, BackupMeta
     }
   }
 
-  public void deleteBackUpInfoById(int id) throws MetaStoreException {
+  public void deleteBackUpInfoById(long id) throws MetaStoreException {
     try {
       backUpInfoDao.delete(id);
     } catch (Exception e) {
